moving HeaderDashboard to ui package
